# internal/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: FabAID Manager API
  description: NSF grant/project tracking — effort, budgets, subawards, invoices, and statements of work.
  version: "1.0.0"
  license:
    name: Apache-2.0
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session cookie or API key (Bearer fabaid_…)
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    BudgetDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        entity_type:
          type: string
        entity_id:
          type: string
          format: uuid
        budget_id:
          type: string
          format: uuid
          nullable: true
        doc_type:
          type: string
        filename:
          type: string
        content_type:
          type: string
        s3_key:
          type: string
        file_size:
          type: integer
          format: int64
        uploaded_by:
          type: string
          nullable: true
        uploaded_by_name:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          nullable: true
...
  parameters:
...
    entityType:
      name: entityType
      in: path
      required: true
      schema:
        type: string
        enum: [grant, subaward]
    entityID:
      name: entityID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    budgetID:
      name: budgetID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /institution-rates/{entityType}/{entityID}/budgets/{budgetID}/documents:
    parameters:
      - $ref: "#/components/parameters/entityType"
      - $ref: "#/components/parameters/entityID"
      - $ref: "#/components/parameters/budgetID"
    get:
      tags: [institution-rates]
      summary: List documents attached to a budget
      operationId: listBudgetDocuments
      responses:
        "200":
          description: Array of budget documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetDocument"
    post:
      tags: [institution-rates]
      summary: Upload a document and attach it to a budget
      operationId: uploadBudgetDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, doc_type]
              properties:
                file:
                  type: string
                  format: binary
                doc_type:
                  type: string
                  description: "Document type (e.g. justification, quote)"
                description:
                  type: string
      responses:
        "201":
          description: Uploaded budget document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDocument"
  /institution-rates/{entityType}/{entityID}/budgets/{budgetID}/documents/{docID}:
    parameters:
      - $ref: "#/components/parameters/entityType"
      - $ref: "#/components/parameters/entityID"
      - $ref: "#/components/parameters/budgetID"
      - name: docID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [institution-rates]
      summary: Get budget document metadata
      operationId: getBudgetDocument
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDocument"
    delete:
      tags: [institution-rates]
      summary: Soft-delete a budget document
      operationId: deleteBudgetDocument
      responses:
        "204":
          description: Deleted
  /institution-rates/{entityType}/{entityID}/budgets/{budgetID}/documents/{docID}/download:
    parameters:
      - $ref: "#/components/parameters/entityType"
      - $ref: "#/components/parameters/entityID"
      - $ref: "#/components/parameters/budgetID"
      - name: docID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [institution-rates]
      summary: Download a budget document file
      operationId: downloadBudgetDocument
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ── Authentication ──────────────────────────────────────
...